+++ datainsight_agent/components/pipeline/__init__.py
@@ -1,6 +1,0 @@
-from .service import SimplePipeline
-
-__all__ = ["SimplePipeline"]
-
-
-