+++ datainsight_agent/components/vector_retriever/__init__.py
@@ -1,6 +1,0 @@
-from .service import VectorRetriever
-
-__all__ = ["VectorRetriever"]
-
-
-